[STD-33] Engine refactoring: systems - Moving to service things related to systems - Organizing engine - Optimizing imports
