+++ webview/engine/core/instances/Texture.ts
@@ -1,4 +1,3 @@
-import IMAGE_WORKER_ACTIONS from "../static/IMAGE_WORKER_ACTIONS"
 import TEXTURE_WRAPPING from "../static/texture/TEXTURE_WRAPPING"
 import TEXTURE_FILTERING from "../static/texture/TEXTURE_FILTERING"
 import TEXTURE_FORMATS from "../static/texture/TEXTURE_FORMATS"
