+++ webview/engine/core/lib/rendering/UIAPI.ts
@@ -1,4 +1,3 @@
-import Engine from "../../Engine"
 import InputEventsAPI from "../utils/InputEventsAPI"
 import QueryAPI from "../utils/QueryAPI"
 import FileSystemAPI from "../utils/FileSystemAPI"
