+++ webview/engine/core/lib/utils/ScriptsAPI.ts
@@ -10,7 +10,6 @@ import CameraAPI from "./CameraAPI"
 import QueryAPI from "./QueryAPI"
 import FileSystemAPI from "./FileSystemAPI"
 import EntityAPI from "./EntityAPI"
-import Engine from "../../Engine"
 import ProjectionEngine from "@lib/ProjectionEngine";
 
 export default class ScriptsAPI {
