+++ webview/engine/core/runtime/BloomSystem.ts
@@ -1,5 +1,3 @@
-import GPU from "../GPU"
-import CameraAPI from "../lib/utils/CameraAPI"
 import StaticFBO from "../lib/StaticFBO"
 import StaticShaders from "../lib/StaticShaders"
 import StaticMeshes from "../lib/StaticMeshes"
@@ -7,7 +5,6 @@ import Framebuffer from "../instances/Framebuffer"
 import MetricsController from "../lib/utils/MetricsController"
 import METRICS_FLAGS from "../static/METRICS_FLAGS"
 import GPUUtil from "../utils/GPUUtil";
-import Engine from "../Engine";
 import ProjectionEngine from "@lib/ProjectionEngine";
 import IEngineSystem from "@engine-core/IEngineSystem";
