+++ webview/engine/core/runtime/renderers/AtmosphereRenderer.ts
@@ -6,8 +6,6 @@ import MetricsController from "../../lib/utils/MetricsController"
 import METRICS_FLAGS from "../../static/METRICS_FLAGS"
 import AtmosphereComponent from "../../instances/components/AtmosphereComponent"
 import {mat4} from "gl-matrix"
-import CameraAPI from "../../lib/utils/CameraAPI"
-import Engine from "../../Engine";
 import ProjectionEngine from "@lib/ProjectionEngine";
 
 const resources = mat4.create().fill(0)
