+++ webview/engine/core/runtime/renderers/SceneRenderer.ts
@@ -2,8 +2,6 @@ import GPU from "../../GPU"
 import StaticMeshes from "../../lib/StaticMeshes"
 import type Entity from "../../instances/Entity"
 import Shader from "../../instances/Shader"
-import Engine from "../../Engine"
-import CameraAPI from "../../lib/utils/CameraAPI"
 import StaticFBO from "../../lib/StaticFBO"
 import PointShadowsSystem from "../PointShadowsSystem"
 import UberMaterialAttributeGroup from "../../lib/UberMaterialAttributeGroup"
