+++ webview/engine/tools/EngineTools.ts
@@ -2,7 +2,6 @@ import GridSystem from "./icons/GridSystem"
 import IconsSystem from "./icons/IconsSystem"
 import GizmoSystem from "./gizmo/GizmoSystem"
 import SelectedSystem from "./outline/SelectedSystem"
-import Engine from "../core/Engine"
 import CameraTracker from "./utils/CameraTracker"
 import WireframeRenderer from "./outline/WireframeRenderer"
 import ENVIRONMENT from "../core/static/ENVIRONMENT"
