+++ webview/engine/tools/icons/IconsSystem.ts
@@ -1,6 +1,4 @@
-import Engine from "../../core/Engine"
 import GPU from "../../core/GPU"
-import CameraAPI from "../../core/lib/utils/CameraAPI"
 import LIGHT_TYPES from "../../core/static/LIGHT_TYPES"
 import LineRenderer from "./LineRenderer"
 import StaticMeshes from "../../core/lib/StaticMeshes"
