+++ webview/engine/tools/outline/WireframeRenderer.ts
@@ -3,7 +3,6 @@ import COLLISION_TYPES from "../../core/static/COLLISION_TYPES"
 import {mat4, vec3} from "gl-matrix"
 import StaticMeshes from "../../core/lib/StaticMeshes"
 import StaticEditorShaders from "../utils/StaticEditorShaders"
-import Engine from "../../core/Engine"
 import StaticFBO from "../../core/lib/StaticFBO"
 import StaticEditorMeshes from "../utils/StaticEditorMeshes"
 import ResourceEntityMapper from "../../core/lib/ResourceEntityMapper"
