+++ webview/engine/tools/utils/CameraTracker.ts
@@ -1,8 +1,6 @@
-import CameraAPI from "../../core/lib/utils/CameraAPI"
 import {quat, vec4} from "gl-matrix"
 import CAMERA_ROTATIONS from "../static/CAMERA_ROTATIONS"
 import GPU from "../../core/GPU"
-import Engine from "../../core/Engine";
 import ProjectionEngine from "@lib/ProjectionEngine";
 
 let holding = false
