+++ webview/window/editor/util/SceneEditorUtil.ts
@@ -3,11 +3,9 @@ import LocalizationEN from "@enums/LocalizationEN"
 import ConversionAPI from "@engine-core/lib/math/ConversionAPI"
 import GPU from "@engine-core/GPU"
 import PickingAPI from "@engine-core/lib/utils/PickingAPI"
-import Engine from "@engine-core/Engine"
 import DepthPrePassSystem from "@engine-core/runtime/DepthPrePassSystem"
 import EngineTools from "../../../engine/tools/EngineTools"
 import {glMatrix, quat} from "gl-matrix"
-import CameraAPI from "@engine-core/lib/utils/CameraAPI"
 import CameraTracker from "../../../engine/tools/utils/CameraTracker"
 import EngineResourceLoaderService from "@services/EngineResourceLoaderService"
 import getViewportContext from "../templates/get-viewport-context"
